renamed playerController file to trackController and update. add insert statements to seeds.sql
