+++ db/seeds.sql
@@ -1,0 +1,7 @@
+use track_db;
+ 
+insert into tracks (name,address,city,state,zipCode,trackType)
+values ('test track 1','1313 Mockingird Lane','Marietta','GA',30060,'standard'),
+ ('test track 2','1111 Main Peachtree Blvd','Kennesaw','GA','30144','standard'),
+ ('test track 3','5555 Atlernate Peachtree Blvd','Kennesaw','GA','30144','standard');
+
